+++ src/rpc/search-leader.hh
@@ -5,23 +5,23 @@
 class SearchLeader : public RPC
 {
 public:
-    SearchLeader(int old_leader);
+    SearchLeader(size_t old_leader);
     SearchLeader(const json& serialized_json);
     SearchLeader(const std::string& serialized);
 
     json serialize_message() const override;
 
-    const int old_leader_;
+    const size_t old_leader_;
 };
 
 class SearchLeaderResponse : public RPC
 {
 public:
-    SearchLeaderResponse(int new_leader);
+    SearchLeaderResponse(size_t new_leader);
     SearchLeaderResponse(const json& serialized_json);
     SearchLeaderResponse(const std::string& serialized);
 
     json serialize_message() const override;
 
-    const int new_leader_;
+    const size_t new_leader_;
 };
